Add values.py processor for environmental overrides of subcharts
This script allows you to have an single environmental YAML file that can be
chunked up for feeding into various subcharts.  This is required as values
input into helm requires the top level namespace be stripped off.  In other
words, a flat namespace.  This allows a single environmental definition that
can be chunked into YAML files for subchart input.

Usage will be documented on the wiki.