Commit neutron chart
This is a functioning neutron chart that leverages a flat network
but supports vxlan and gre networks in values.yaml.  We were unable
to test the vxlan functionality as it tickles the bnx2x cards in our
HP blade lab.

For now, this leverages daemonsets which make sense.  It does require
a new label, namely openvswitch=enabled as we need a label that can be
applied both to the control plane and tenant compute hosts as both
require neutron agents and openvswitch.  The interfaces today match
our labs, namely enp11s0f0 for the flat network on physnet1 and
enp12s0f0 for the external network.  These can be overriden in
values.yaml via set or a global environmental file.

It depends on the keystone endpoint work.

This chart was tested against a working nova chart not commited as
DTadrzak has one open in PR#45