Add travis-ci to openstack-helm
Adds the travis.yml file to openstack-helm. The current travis
workflow is: install glide for helm, build helm from source,
run kubeadm in a container, init helm on the client side, then run
tiller locally. The scripts currently run the makefile to
check: helm dep up, helm lint, and helm package on the charts
defined in the makefile. Then a helm install --dry-run debug is
run on each of the charts packaged locally.