Remove Personal Copywrite from DB Script
I am removing my copywrite and transfering it to the OpenStack-Helm Authors
for the DB Managment Script. As although this was primarily written while an
independant OSS Developer, it is not currently present in any other codebase
and will only be merged post commencing work on the project in an official
capacity. This does not affect the copywrite of any other code contibuted to
the project by myself or any other party.