add node type option to rabbitmq clustering
Currently, rabbitmq clustering is using the autocluster plugin,
and NODE_TYPE is set to disc in default. so every nodes join cluster as disc node.
However, since there is a need for disc + ram clustering for the performance.
if change node_type from disc to ram, clustering configured as [disc + ram + ram].

Change-Id: Ie83689b0554f0f993bdffac666f0f56db8082992