Remove ConfigMap hash annotations from StatefulSets
Kubernetes does not allow updating annotations on StatefulSets (it
actually only allows a couple of fields to be modified), so the
existence of these hash annotations makes it impossible to modify
configmap values.

This will be a problem not only for upgrades, but for normal operation
with charts like `mariadb`, which will occasionally require
`helm upgrade ...` to modify settings to allow recovery from certain
cluster failure modes.

Change-Id: I5f0433fc05c5aa22b4ad7ec296efbde3acca3d5a