Clarified a couple documentation points.
Conditionals:  since the developer flag is no longer used by the
glance chart, it's no longer a good example.  However, we want to
keep the flag around for future use if needed.  So, replaced with
a foo/bar example.

gen-oslo-openstack-helm README:  it referred to "this directory",
which doesn't make sense when the README is sourced into the
overall ops guide.  Specified the directory explicitly.

Change-Id: I2ba02bd3efd87e58fcb701c8b7c2bb6af05994ba