Remove admin_token_auth
The admin_token_auth in keystone paste posts a security issue, and has
been deprecated in the M release, and removed in O release.  Operators
should be using ``keystone-manage bootstrap`` to bootstrap the "admin"
user.  This patch set removes the filter and its usage in the various
pipelines from the paste.ini file.

Implements: bp keystone-admin-token-disable
External-Tracking-Id: OSH-102

[0] https://blueprints.launchpad.net/openstack-helm/+spec/keystone-admin-token-disable

Change-Id: I5ae29cad4f7daa4bc8fa117b6f5ed998c2ec6cad