Horizon: move to kolla 4.0.0 image
This PS moves the default horizon image to Kolla 4.0.0 - it is
backwards compatible with the 3.0.0 apis. The reason for changing is
the image iteslf is much enhanced with plugins installed for many
services: murano, designate etc that we deploy with openstack-helm.

The alternative would be producing a custom image for the project
if we wish to support these services.

Change-Id: Ib71cd27f077dd94a7ab56e621d13f0ba4ffb4e49