Change the documentation table format
As we update the version information in the document, the ascii rst
table easily gets misaligned and causes the doc gate to fail.  This
patch set changes the table structure, while retaining the formatting,
so future update of the document can be done easily without spending
time realigning the table cell border.

Change-Id: I84aeba3604ced255b057b7ce43f1f94a1af19b65