Remove /var/cache/cinder emptyDir volume
This patch removes unused `/var/cache/cinder` volume that is mounted for
each of Cinder services. I suppose it was added to serve as value of
`signing_dir` option from keystonemiddleware and is used to cache PKI
tokens. We're however never setting the option to that value and the
option was deprecated in Ocata, so we should be fine to remove it.

In case someone still wanted to use `signing_dir` config option, he can
easily confiure it back using config and volumes overrides.

Related-Bug: 1702072

Change-Id: I03460151e01293b19a864528561c80cab1a72114