Fix ceph single node gate health
Ceph's health when running in the single node gate currently
is 'HEALTH_WARN'.  This PS adds an extra config option such that
the ceph cluster will not attempt to perform replication when deployed
on a single node, as well as introduces a mvp yaml file for ceph
deployments in the gate.

Change-Id: Ib2ec3345140f541c94da044ff9d77723ea3ee2bd