Gate: Loopback device support
This PS adds loopback device support to the gate scripts.
Rather than using simple loopback devices we use iscsi to
allow is to target the created devices via the bus they are
connected to. An arbitary number of devices of a desired size
can be created, and controlled via env vars.

Change-Id: I05fb7f3a1564bc36903aed2c46ed996bb8cc57c8