Nova: Update bootstrap job
This PS updates the nova boostrap job, both bringing it inline with
other services and also adding the default flavours from Mitaka.
By setting the m1.tiny size to match the upstream it is expected
that the majority of gate failures will be solved - as many of the
VMs in nodepool do not have 20GB of free space on the rootfs.

Change-Id: I767795015422a1b326a6f34b5a3f50bf41a0ca4b