Ceph: Fix Multinode Gates
This PS should make the multinode gate more reliable for ceph.

It's not perfect in that is uses the default route IPs in zuul, which
has a performance hit. The bash is also ugly as sin - but it gets us
moving again until somone has time to implement a more eligant solution
that works both in infra supplied nodes and other CI systems.

Change-Id: Ie1326f048c43d04a945ca155bde8824d17ddcbec