Clarify type of auto_bridge_add
With auto_bridge_add defined as:

  auto_bridge_add:
  # nothing here

helm doesn't treat is as am empty table so issues warnings such as:

    2017/10/03 19:24:37 warning: destination for auto_bridge_add is a table. Ignoring non-table value <nil>

This happens most of the time as we use `helm --values=...` to augment
values.yaml in the chart.

This change clarifies it is indeed a table and the warnings no long
appear.

Change-Id: Ib733ed72df49e134a9199e4024b2893629b990dc