Nova: Use endpoint lookups for NoVNC proxy
This PS updates nova to use endpoint lookups for the nova VNC proxy.
This enables the proxy to be accessed from outside the openstack
namespace and the cluster via its FQDN and also brings it inline
with other services.

Addtionally some spacing issues are resolved in the nova etc configmap
to make them consistent across all charts.

Change-Id: Ie93420d6a19915bee2750fbe7fd4c21fc0641ee8