Horizon: move to kolla image with bugs fixed
Horizon has used the Kolla 4.0.0 image for some time as the 3.0.3
image pushed to dockerhub had several bugs in it. The 4.0.0 image
also has a bug where the topology view is not enabled for heat.
The 4.0.0 image is based on ocata, by moving to the `ocata` tag
we can use a image where this bug has been fixed.

Change-Id: Icca1563561454c92fcebd61f14e08f1d19289d24