Fix elasticsearch repository configuration entry
This dynamically adds the elasticsearch path.repo configuration
entry if it's not defined.  This solves issues arising when the
storage settings are disabled in favor of emptydirs for simpler
ES deployments.  If elasticsearch attempts to configure the repo
path with an invalid entry (inaccesible external or shared fs
path), the service will crash.

Change-Id: I089b77104107dfb1f8e6ea2d8a560384718e63f9