Reduce doc max depth
Current setting the maxdepth of 3 causes all the sub-headings
(e.g. sub-titles under specs) to be displayed, needless expanding
the table of content making the doc hard to read. This patch set
reduces the maxdepth to 2 to reduce the noise in the OpenStack-Helm
documentation.

Also, this patch set touches up on RST structure and capitalization
changes in the document.

Change-Id: I7b8845f42180eafe08ed68a6c5a1e82ea49927be
Signed-off-by: Tin Lam <tin@irrational.io>