mariadb: remove connect_timeout to fix mysqldump
Remove 'connect_timeout' which defined in the '[client]' section which
surprisingly affects mysqldump by causing it to fail as it doesn't
understand this; we're not doing anything that depends on having this
value so this is safe.

This now means mysqldump works (as expected) and we can use the
packaged/pod version of mysqldump for backup purposes.

Change-Id: I9142969e4b8576b19353c063586dd11f91875f0b
LocalWords:  mysqldump