Fixes gnocchi with postgresql deployment if running on different hosts
a) The db-init method ignores the configured admin password.

This seems to work if db-init was scheduled to the same host as
postgresql is running on but fails on a remote connection. This change
sets the PGPASSWORD variable to the provided admin password.

b) Fixed liveness and readiness probe for postgresql

While this seems to work it continously outputs the message
"psql: FATAL: role "root" does not exist"

This change adds the admin username as argument to pg_isready
to stop this message.

Update: Making the linter happy ;-)

Change-Id: I59b47a8724065d08203c29172c15350aaa785203
Closes-Bug: #1749177