Allows the configuration of SSO choices for horizon
This change allows enabling the WEBSSO login screen on horizon,
which allows to choose from one or more configured SSO providers.

Example configuration

 local_settings:
  auth:
    sso:
      enable: true
      initial_choice: "acme_oidc"
    idp_mapping:
      - name: "acme_oidc"
        label: "Acme Corporation - OpenID Connect"
        idp: "myidp1"
        protocol: "oidc"
      - name: "acme_saml2"
        label: "Acme Corporation - SAML2"
        idp: "myidp2"
        protocol: "saml2"

The initial_choice defaults to "credentials" which is the default
Keystone Credential authentication.

The values for idp: and protocol: will be used to construct the redirect
URL for keystone, which will look like:

 /v3/OS-FEDERATION/identity_providers/<idp>/protocols/<protocol>/auth

Change-Id: I44e11880292176114753274f965bcd0c2cd01302