Excise deprecated or removed LDAP options
Some of the LDAP configurations were deprecated in N- or O-release have
been since removed, namely the option to allow for write operations in
LDAP.  This is causing some issue with more recent releases of keystone
as discovered in [0]. This patch set removes them from the example to
avoid the errors for more recent releases without impacting N- or
O-releases as they were already deprecated at that point.

Also, this adds in a configuration sample that bootstraps an LDAP
domain by default.

[0] https://review.openstack.org/#/c/561751/

Change-Id: Ie8f23cd61dad1860a41422264945c169f0488729
Signed-off-by: Tin Lam <tin@irrational.io>