Update curator auth config and actions
Curator is unable to use environment variables for configuration
values if the configured option contains more than the env
variable.  In the case of the http_auth value (which expects
user:password), using ${USER}:${PASS} prevents curator from
successfully authenticating to elasticsearch.  This moves to
dynamically define these values in the configmap if the value is
empty

This also updates values for curators actions to target logstash-
indices for its actions

Change-Id: Id5b49171e00847432e4ab0cf4be60005b70c21e3