Update Makefile
Per [0], update Makefile to behave as it was before with an added
check(s) for the existence of a folder prior to processing.  This
allows us to reinstate the helm-toolkit folder even though it has
been migrated, and will build as usual if a developer has a helm-
toolkit in the directory.

Also, did some minor cleanup to the Makefile.

[0] https://review.openstack.org/#/c/568177/1

Change-Id: I4e0a26b4a228d63a1c769d2f0c935fa3fc56192c
Signed-off-by: Tin Lam <tin@irrational.io>