Armada manifest updates for OSH
This updates the armada multinode manifest to include the new
format for waiting on releases to complete. This is now configured
to wait for a release to complete by checking resources with the
release_group label for a given chart.

This also adds the LDAP chart, which is required for the LMA
services

Change-Id: Ie23073a04dc4c55291c0314beb0d8d215005d353