Update prometheus rule for terminated containers in pods
This updates the prometheus rule for checking for terminated
containers in pods. The previous rule checked for any terminations,
which raised alarms due to completed containers in jobs
being included, which isn't desired behavior. This changes the
expression to check for any containers that have terminated with
a status other than completed

Change-Id: I88e533a56f81f81bd1a81420ecfb7d43ac9e2d0b