Update nfs-provisioner and ldap deployment scripts
This updates the lma-nfs-provisioner deployment script to run make
on the nfs-provisioner chart. Previously, it relied on the local
image registry deployment step being run to lint and package the
nfs-provisioner chart. This change allows the nfs-provisioner to
be deployed without the local image registry if desired.

This also updates the ldap deployment script to only run make on
the ldap chart instead of pulling the default images with the
make-pull directive

Change-Id: I3d8f321d7a8af50fd80ffbd6a337fa17675f5700