Fix pip package version
The python "cmd2" package installs version 0.9+ on Python 2.7, which only
support Python 3.4+. This causes a dependency error in the gate due to
issue outlined in [0].  We will pre-install a capped version of "cmd2"
that would work with Python 2.7.

This patch set also temporarily make the keystone webhook non-voting as
this bug is causing a circular dependencies with patch [1]. Once this is
fixed, a new patch will be submitted to make this voting again.

[0] https://github.com/python-cmd2/cmd2/issues/421
[1] https://review.openstack.org/#/c/571093/

Change-Id: I34cbde65b74efc2805bd7785f84878783d2badbf