Kube-State-Metrics: Change default image used
This changes the default image for kube-state-metrics to use the
bitnami image instead of the coreos image. This allows us to
override the image entrypoint, as the Alpine based image used
previously did not easily allow us to do so. Adding this also
makes creating a common prometheus exporter deployment template
easier, as it reduces the functional differences between exporter
charts and templates

Change-Id: I6c4aac36f563fcb15f52640bc6f9913b45b4358a