Add fluentbit sidecars to ceph mon and ceph osd charts
This adds fluentbit sidecar containers to the ceph mon and ceph
osd charts, allowing for the gathering of ceph logs that arent
sent to stdout and stderr. This sidecar container mounts a shared
directory on /var/log/ceph, and fluentbit is configured to tail
the log files in this directory and tag the event messages
appropriately before forwarding them to fluentd. This behavior can
be toggled on or off via values, and its proposed to be disabled
by default

Change-Id: Ia2c944a68f9e7d831ac0fe11ed09468d0dc9ce71