Grafana: Provision data sources via dynamic template in values
This moves to define the datasources provisioned by grafana via
a template defined in the values.yaml. This allows us to define
multiple datasource types that can be mapped directly to the
corresponding entries in endpoints, which enables us to generate
the data source urls via endpoint lookups rather than hardcoding
this. This is the first step to support multiple data sources in
a singular grafana deployment

Change-Id: Iac7f4b1e07aaf83ae4d2a0c923cd06817f0d8c0d