Grafana: Update LDAP configuration, update volume mounts
This updates the LDAP configuration for grafana, using a template
defined in the values.yaml file. Using the template allows us to
dynamically define LDAP configuration values, such as the bind dn,
search base and group search base paths, the password, and the
LDAP fqdn.  This also updates the volume mount for the
provisioning directory to be defined by the configuration value in
the values.yaml file

Change-Id: I1e4866d1189cf40b08b3443dc725646a1b76094c