Allow overriding some critical postgre parameters
Ideally, this would completely own the postgresql.conf file that
is autogenerated by the container.  This should not be done without
further deliberation though on what impact that may have. For now,
this allows some critical items to be manipulated at installation
time but does not meet the full need of complete configuration
flexibility.

Change-Id: I94fd3ac42398f8114c0b5c8b4ef0a6a96e530f96