Gate: Fix version of ansible deployed via dev-deploy
This pins the version of ansible deployed via the makefile
dev-deploy directives to the version used in zuul.  This was
causing issues with docker, as make dev-deploy setup-host was
deploying ansible 2.6. Ansible 2.6 introduces a new flag to the
docker_container module (init) that is incompatible with our
current roles, which resulted in observed failures in osh-infra
and osh

Change-Id: Ibc885b53bce77eb36817024b21efb0e99865f690