Add sla check in rally task
The current rally job is always execute report command at the end,
and it always succeeds regardless of the rally task result.
Since the values.yaml in the rally chart already includes sla check,
you should execute sla-check command at the end of the job
so that you can check whether the test has passed or not.

Change-Id: I4c91950b86622b1211aea3e9e908cb04711c0e76