Fluent-logging: Update tests and template job
This updates the helm tests and the elasticsearch template job.
This changes the tests to conditionally check whether the
template job is enabled and the templates key is not empty, and
uses the result to determine whether to test for the existence
of those templates (to account for situations where the job is
disabled).

This updates the job to also check whether there are templates
defined in additional to checking whether the job itself is
enabled.

Change-Id: I14cedeb8d8a4444a73ea974426c3b0f136d1b698