Ceph-Mon/Ceph-OSD: Update default fluentbit configuration
As of 0.12.14, fluentbit exposes a flag for setting the db_sync
behavior for writing the location of the tail input to its sqlite
database. The default setting is Full, which introduces additional
synchronizations before and after a transaction. This has the
potential to negatively affect disk performance with the extra
synchronizations. This moves the setting in the chart to Normal,
which performs fewer synchronizations and still maintains a high
level of safety with status writes

Change-Id: I5de8ee40162a50e1ff3395d0e302974b77dd5f9c