Docs: update proxy troubleshooting guide
Currently, the proxy troubleshooting guide does not include the IP
address of the Kubernetes API in the defined environment or Ansible
playbook variables. This causes deployments to fail when requests are
routed through proxy servers. This change adds a proxy configuration
section to the beginning of the installation guide and includes the
process of adding the Kubernetes API IP address to the environment and
Ansible playbook variables. It also removes the unecessary sections from
the troubleshooting guide and upgrades the note about resolv.conf being
overwritten to a warning.

Change-Id: I1ff753517b387027f0cc052b18e8822b638cfb4d